Fix compiler warnings under linux.

+++ src/eventlog/eventlog.h
@@ -64,7 +64,7 @@ class EVENTLOG_API EventLog : public IEventLog, public EventLogIndex
         ~EventLog();
 
         virtual ProgressMonitor setProgressMonitor(ProgressMonitor newProgressMonitor);
-        virtual void setProgressCallInterval(double seconds) { progressCallInterval = seconds * CLOCKS_PER_SEC; lastProgressCall = clock(); }
+        virtual void setProgressCallInterval(double seconds) { progressCallInterval = (long)(seconds * CLOCKS_PER_SEC); lastProgressCall = clock(); }
         void progress(); // notify monitor
 
         void parseInitializationLogEntries();
